BE/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.4
pydantic-settings==2.7.0
python-multipart==0.0.12
supabase==2.9.1
email-validator==2.1.1
PyJWT[crypto]==2.8.0
python-dotenv==1.0.1
openai==1.67.0
google-genai==1.0.0
anyio>=4.8,<5
anthropic>=0.69.0
Pillow==10.4.0
requests>=2.31.0
svix==1.82.0
httpx==0.27.2
